requirements.txt
blinker==1.8.2
breadability==0.1.20
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docopt==0.6.2
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
language-tool-python==2.8
lxml==5.2.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
pycountry==23.12.11
PyPDF2==3.0.1
python-docx==1.1.2
regex==2024.5.15
requests==2.32.2
setuptools==70.0.0
sumy==0.11.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
